+++ src/h3sed/conf.py
@@ -9,7 +9,7 @@ This file is part of h3sed - Heroes3 Savegame Editor.
 Released under the MIT License.
 
 @created   14.03.2020
-@modified  15.07.2026
+@modified  26.07.2026
 ------------------------------------------------------------------------------
 """
 try: from ConfigParser import RawConfigParser                 # Py2
@@ -26,8 +26,8 @@ import sys
 """Program title, version number and version date."""
 Name = "h3sed"
 Title = "Heroes3 Savegame Editor"
-Version = "3.7.dev4"
-VersionDate = "15.07.2026"
+Version = "3.7.dev5"
+VersionDate = "26.07.2026"
 
 Frozen = getattr(sys, "frozen", False)
 if Frozen: