+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 3.7, 2026-07-13
 ---------------
 - add hero profile tab, with hero faction and biography
+- ensure hero copy-paste not overwriting read-only properties
 
 
 3.6, 2026-05-31