+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 CHANGELOG
 =========
 
-3.7, 2026-07-13
+3.7, 2026-07-26
 ---------------
 - add hero profile tab, with hero faction and biography
 - add hero map coordinates to hero profile (issue #31)
 - ensure hero copy-paste not overwriting read-only properties
+- fix overquoting in CSV export
 
 
 3.6, 2026-05-31